fix more warnings
svn path=/trunk/kdeedu/kvoctrain/; revision=294303

+++ kvoctrain/kvt-core/kvd_wkvtml.cpp
@@ -724,7 +724,7 @@ bool kvoctrainDoc::saveToKvtMl (QTextStream& os, QString &title) {
 
     ent_no++;
     if (ent_percent != 0 && (ent_no % ent_percent) == 0 )
-      emit progressChanged(this, ent_no / f_ent_percent);
+      emit progressChanged(this, ent_no / (int) f_ent_percent);
 
     if ((*first).getLesson() != 0) {   // entry belongs to lesson x
       QString ls;
